version bug fixes, add template search for messages
